pyproject.toml
[project]
name = "alexey-agen"
version = "0.1.0"
description = ""
authors = [
    {name = "TiRe",email = "zientenin@mail.ru"}
]
readme = "README.md"
requires-python = "=3.13.7"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
